Clean up code
Add IsOnFrustum method to Chunk.cs
Remove unnecessary profiling statements.
