+++ source/client/source/UI/Windows/ProfilerWindow.cs
@@ -1,5 +1,6 @@
 ﻿using ImGuiNET;
 using Korpi.Client.Debugging;
+using Korpi.Client.Debugging.Profiling;
 using Korpi.Client.Window;
 
 namespace Korpi.Client.UI.Windows;
